# faster method to calculate PI