--- a/math/cosine.go
+++ b/math/cosine.go
@@ -31,7 +31,7 @@ func prepCosine(z, x *decimal.Big, ctx decimal.Context) (*decimal.Big, int, bool
 		uv := arith.Abs(v)
 
 		// Adjust so we have ceil(v/10) + ctx.Precision, but check for overflows.
-		// 1+((v-1)/10) will be widly incorrect for v == 0, but x/y = 0 iff
+		// 1+((v-1)/10) will be wildly incorrect for v == 0, but x/y = 0 iff
 		// x = 0 and y != 0. In this case, -2pi <= x >= 2pi, so we're fine.
 		prec, ok := checked.Add(1+((uv-1)/10), uint64(ctx.Precision))
 		if !ok || prec > maxInt {