added a faster method to calculate Pi, and added a last used cache to provide an even faster Pi function, added tests and benchmarks, fixed a spelling mistake and added some new consts used in the new pi method

Additional beneficial side effects: trig functions are now faster with precision values in the hundreds (overhead still mostly outweighs benefit at lower precision values)

**Before:**

> BenchmarkPi/5-16  	 3000000	       415 ns/op
> BenchmarkPi/16-16 	 5000000	       361 ns/op
> BenchmarkPi/25-16 	 5000000	       395 ns/op
> BenchmarkPi/32-16 	 5000000	       398 ns/op
> BenchmarkPi/41-16 	 3000000	       441 ns/op
> BenchmarkPi/50-16 	 3000000	       416 ns/op
> BenchmarkPi/75-16 	 3000000	       485 ns/op
> BenchmarkPi/82-16 	10000000	       232 ns/op
> BenchmarkPi/97-16 	10000000	       184 ns/op
> BenchmarkPi/100-16         	50000000	        32.7 ns/op
> BenchmarkPi/137-16         	    3000	    417051 ns/op
> BenchmarkPi/250-16         	    2000	   1022068 ns/op
> BenchmarkPi/333-16         	    1000	   1642012 ns/op
> BenchmarkPi/500-16         	     500	   3328571 ns/op
> BenchmarkPi/646-16         	     300	   5890675 ns/op
> BenchmarkPi/750-16         	     200	   7617458 ns/op
> BenchmarkPi/943-16         	     100	  13133408 ns/op
> BenchmarkPi/1500-16        	      50	  32773017 ns/op
> BenchmarkPi/5000-16        	       2	 621706059 ns/op
> BenchmarkPi/7500-16        	       1	1864888228 ns/op
> BenchmarkPi/15000-16       	       1	12172289166 ns/op
> 

> BenchmarkCos/0/5-16         	  200000	      8969 ns/op
> BenchmarkCos/0/16-16        	  200000	     10073 ns/op
> BenchmarkCos/0/25-16        	  200000	     11145 ns/op
> BenchmarkCos/0/32-16        	  200000	     11089 ns/op
> BenchmarkCos/0/41-16        	  200000	     11484 ns/op
> BenchmarkCos/0/50-16        	  200000	     11280 ns/op
> BenchmarkCos/0/75-16        	  100000	     11787 ns/op
> BenchmarkCos/0/82-16        	  100000	     11829 ns/op
> BenchmarkCos/0/97-16        	  100000	     12041 ns/op
> BenchmarkCos/0/100-16       	    5000	    296543 ns/op
> BenchmarkCos/0/137-16       	    3000	    446739 ns/op
> BenchmarkCos/0/250-16       	    2000	   1064858 ns/op
> BenchmarkCos/0/333-16       	    1000	   1693124 ns/op
> BenchmarkCos/0/500-16       	     500	   3449743 ns/op
> BenchmarkCos/0/646-16       	     200	   5973578 ns/op
> BenchmarkCos/0/750-16       	     200	   8058216 ns/op
> BenchmarkCos/0/943-16       	     100	  13791639 ns/op
> BenchmarkCos/0/1500-16      	      50	  32450864 ns/op
> BenchmarkCos/0/5000-16      	       2	 618845580 ns/op
> BenchmarkCos/0/7500-16      	       1	1839998816 ns/op
> BenchmarkCos/0/15000-16     	       1	12045381051 ns/op
> ...
> BenchmarkCos/8/5-16         	   10000	    179060 ns/op
> BenchmarkCos/8/16-16        	   50000	     26049 ns/op
> BenchmarkCos/8/25-16        	   50000	     34323 ns/op
> BenchmarkCos/8/32-16        	   20000	     62587 ns/op
> BenchmarkCos/8/41-16        	   20000	     64984 ns/op
> BenchmarkCos/8/50-16        	   50000	     38073 ns/op
> BenchmarkCos/8/75-16        	   20000	     70869 ns/op
> BenchmarkCos/8/82-16        	   30000	     42685 ns/op
> BenchmarkCos/8/97-16        	   20000	     75581 ns/op
> BenchmarkCos/8/100-16       	    5000	    303128 ns/op
> BenchmarkCos/8/137-16       	    3000	    499208 ns/op
> BenchmarkCos/8/250-16       	    2000	   1000899 ns/op
> BenchmarkCos/8/333-16       	    1000	   1534902 ns/op
> BenchmarkCos/8/500-16       	     500	   3002906 ns/op
> BenchmarkCos/8/646-16       	     300	   4801711 ns/op
> BenchmarkCos/8/750-16       	     200	   7299330 ns/op
> BenchmarkCos/8/943-16       	     100	  12052573 ns/op
> BenchmarkCos/8/1500-16      	      50	  29603079 ns/op
> BenchmarkCos/8/5000-16      	       2	 601708068 ns/op
> BenchmarkCos/8/7500-16      	       1	1841416724 ns/op
> BenchmarkCos/8/15000-16     	       1	12092185757 ns/op

**After:**

> BenchmarkPi/5-16  	100000000	        10.2 ns/op
> BenchmarkPi/16-16 	100000000	        10.0 ns/op
> BenchmarkPi/25-16 	100000000	        10.1 ns/op
> BenchmarkPi/32-16 	100000000	        10.2 ns/op
> BenchmarkPi/41-16 	100000000	        10.1 ns/op
> BenchmarkPi/50-16 	100000000	        10.0 ns/op
> BenchmarkPi/75-16 	100000000	        10.0 ns/op
> BenchmarkPi/82-16 	100000000	        10.1 ns/op
> BenchmarkPi/97-16 	100000000	        10.2 ns/op
> BenchmarkPi/100-16         	100000000	        10.2 ns/op
> BenchmarkPi/137-16         	100000000	        10.1 ns/op
> BenchmarkPi/250-16         	100000000	        10.0 ns/op
> BenchmarkPi/333-16         	100000000	        10.0 ns/op
> BenchmarkPi/500-16         	100000000	        10.1 ns/op
> BenchmarkPi/646-16         	100000000	        10.1 ns/op
> BenchmarkPi/750-16         	100000000	        10.0 ns/op
> BenchmarkPi/943-16         	100000000	        10.1 ns/op
> BenchmarkPi/1500-16        	100000000	        10.3 ns/op
> BenchmarkPi/5000-16        	100000000	        10.1 ns/op
> BenchmarkPi/7500-16        	200000000	        10.1 ns/op
> BenchmarkPi/15000-16       	200000000	        10.1 ns/op

> BenchmarkCos/0/5-16         	  200000	      8197 ns/op
> BenchmarkCos/0/16-16        	  200000	      9654 ns/op
> BenchmarkCos/0/25-16        	  200000	     10954 ns/op
> BenchmarkCos/0/32-16        	  200000	     10949 ns/op
> BenchmarkCos/0/41-16        	  200000	     11141 ns/op
> BenchmarkCos/0/50-16        	  200000	     11070 ns/op
> BenchmarkCos/0/75-16        	  100000	     11699 ns/op
> BenchmarkCos/0/82-16        	  200000	     11855 ns/op
> BenchmarkCos/0/97-16        	  100000	     12397 ns/op
> BenchmarkCos/0/100-16       	  200000	     12303 ns/op
> BenchmarkCos/0/137-16       	  100000	     13581 ns/op
> BenchmarkCos/0/250-16       	  100000	     18318 ns/op
> BenchmarkCos/0/333-16       	  100000	     22951 ns/op
> BenchmarkCos/0/500-16       	   50000	     35841 ns/op
> BenchmarkCos/0/646-16       	   30000	     49796 ns/op
> BenchmarkCos/0/750-16       	   20000	     64784 ns/op
> BenchmarkCos/0/943-16       	   20000	     91778 ns/op
> BenchmarkCos/0/1500-16      	   10000	    200966 ns/op
> BenchmarkCos/0/5000-16      	    1000	   1901300 ns/op
> BenchmarkCos/0/7500-16      	     300	   4213997 ns/op
> BenchmarkCos/0/15000-16     	     100	  16530330 ns/op
> ...
> BenchmarkCos/8/5-16         	   10000	    181458 ns/op
> BenchmarkCos/8/16-16        	   10000	    189211 ns/op
> BenchmarkCos/8/25-16        	   20000	     62073 ns/op
> BenchmarkCos/8/32-16        	   20000	     61744 ns/op
> BenchmarkCos/8/41-16        	   20000	     64763 ns/op
> BenchmarkCos/8/50-16        	   10000	    166330 ns/op
> BenchmarkCos/8/75-16        	   20000	     70651 ns/op
> BenchmarkCos/8/82-16        	   20000	     73841 ns/op
> BenchmarkCos/8/97-16        	   20000	     74366 ns/op
> BenchmarkCos/8/100-16       	   30000	     48158 ns/op
> BenchmarkCos/8/137-16       	   30000	     52445 ns/op
> BenchmarkCos/8/250-16       	   20000	     76605 ns/op
> BenchmarkCos/8/333-16       	   20000	     95718 ns/op
> BenchmarkCos/8/500-16       	   10000	    139082 ns/op
> BenchmarkCos/8/646-16       	   10000	    180516 ns/op
> BenchmarkCos/8/750-16       	   10000	    198099 ns/op
> BenchmarkCos/8/943-16       	   10000	    160323 ns/op
> BenchmarkCos/8/1500-16      	    3000	    538973 ns/op
> BenchmarkCos/8/5000-16      	     500	   3616287 ns/op
> BenchmarkCos/8/7500-16      	     100	  17633691 ns/op
> BenchmarkCos/8/15000-16     	      10	 119275707 ns/op