@nathanhack 2019-07-28
Ok so after a bit of deliberation, I've decided I don't like any of the implementations using locks that I tried (maybe later if someone complains about it we can do something to help at that time). So I have removed the Pi cache part. So i've only swapped out the continuing fractions form for piChudnovskyBrothers(). Since anything under 100 precision we use the const, this should provide some significant speedups for anyone using it for bigger precision sizes.

BenchmarkPiContinuedFrac/5-16 500000	2301 ns/op
BenchmarkPiContinuedFrac/6-16 500000	2662 ns/op
BenchmarkPiContinuedFrac/7-16 300000	3432 ns/op
BenchmarkPiContinuedFrac/8-16 300000	3770 ns/op
BenchmarkPiContinuedFrac/9-16 300000	4667 ns/op
BenchmarkPiContinuedFrac/10-16 300000	5201 ns/op
BenchmarkPiContinuedFrac/11-16 200000	6829 ns/op
BenchmarkPiContinuedFrac/12-16 200000	7470 ns/op
BenchmarkPiContinuedFrac/13-16 200000	8971 ns/op
BenchmarkPiContinuedFrac/14-16 200000	9684 ns/op
BenchmarkPiContinuedFrac/15-16 200000	11007 ns/op
BenchmarkPiContinuedFrac/16-16 100000	12767 ns/op
BenchmarkPiContinuedFrac/25-16 30000	40834 ns/op
BenchmarkPiContinuedFrac/32-16 30000	55971 ns/op
BenchmarkPiContinuedFrac/41-16 20000	79570 ns/op
BenchmarkPiContinuedFrac/50-16 10000	104877 ns/op
BenchmarkPiContinuedFrac/75-16 10000	180373 ns/op
BenchmarkPiContinuedFrac/82-16 10000	202016 ns/op
BenchmarkPiContinuedFrac/97-16 5000	269603 ns/op
BenchmarkPiContinuedFrac/100-16 5000	266486 ns/op
BenchmarkPiContinuedFrac/137-16 5000	385980 ns/op
BenchmarkPiContinuedFrac/250-16 2000	1089872 ns/op
BenchmarkPiContinuedFrac/333-16 1000	1795716 ns/op
BenchmarkPiContinuedFrac/500-16 500	3574321 ns/op
BenchmarkPiContinuedFrac/646-16 200	6868965 ns/op
BenchmarkPiContinuedFrac/750-16 200	9044343 ns/op
BenchmarkPiContinuedFrac/943-16 100	16091890 ns/op
BenchmarkPiContinuedFrac/1500-16 30	36687125 ns/op
BenchmarkPiContinuedFrac/5000-16 2	610959027 ns/op
BenchmarkPiContinuedFrac/7500-16 1	1788974484 ns/op
BenchmarkPiContinuedFrac/15000-16 1	11517741370 ns/op

BenchmarkPiChudnovskyBrothers/5-16 200000	6597 ns/op
BenchmarkPiChudnovskyBrothers/6-16 200000	7256 ns/op
BenchmarkPiChudnovskyBrothers/7-16 200000	7145 ns/op
BenchmarkPiChudnovskyBrothers/8-16 200000	6696 ns/op
BenchmarkPiChudnovskyBrothers/9-16 200000	7202 ns/op
BenchmarkPiChudnovskyBrothers/10-16 200000	6650 ns/op
BenchmarkPiChudnovskyBrothers/11-16 200000	6952 ns/op
BenchmarkPiChudnovskyBrothers/12-16 200000	8863 ns/op
BenchmarkPiChudnovskyBrothers/13-16 200000	8429 ns/op
BenchmarkPiChudnovskyBrothers/14-16 200000	9456 ns/op
BenchmarkPiChudnovskyBrothers/15-16 200000	10067 ns/op
BenchmarkPiChudnovskyBrothers/16-16 200000	9973 ns/op
BenchmarkPiChudnovskyBrothers/25-16 200000	9163 ns/op
BenchmarkPiChudnovskyBrothers/32-16 200000	11273 ns/op
BenchmarkPiChudnovskyBrothers/41-16 100000	12885 ns/op
BenchmarkPiChudnovskyBrothers/50-16 100000	13671 ns/op
BenchmarkPiChudnovskyBrothers/75-16 100000	18122 ns/op
BenchmarkPiChudnovskyBrothers/82-16 100000	19490 ns/op
BenchmarkPiChudnovskyBrothers/97-16 100000	22131 ns/op
BenchmarkPiChudnovskyBrothers/100-16 100000	21690 ns/op
BenchmarkPiChudnovskyBrothers/137-16 50000	26556 ns/op
BenchmarkPiChudnovskyBrothers/250-16 30000	55915 ns/op
BenchmarkPiChudnovskyBrothers/333-16 20000	65747 ns/op
BenchmarkPiChudnovskyBrothers/500-16 10000	107226 ns/op
BenchmarkPiChudnovskyBrothers/646-16 10000	140266 ns/op
BenchmarkPiChudnovskyBrothers/750-16 10000	158381 ns/op
BenchmarkPiChudnovskyBrothers/943-16 10000	219730 ns/op
BenchmarkPiChudnovskyBrothers/1500-16 5000	386462 ns/op
BenchmarkPiChudnovskyBrothers/5000-16 1000	1194042 ns/op
BenchmarkPiChudnovskyBrothers/7500-16 1000	1968074 ns/op
BenchmarkPiChudnovskyBrothers/15000-16 200	6001494 ns/op